make custom_ui as default ui and advanced_user to False
