+++ facefusion/processors/modules/age_modifier.py
@@ -284,6 +284,46 @@ def modify_age(target_face : Face, temp_vision_frame : VisionFrame) -> VisionFra
 		# Convert to final output format
 		paste_vision_frame = cv2.cvtColor((image * 255).astype(np.uint8), cv2.COLOR_RGB2BGR)  # (H, W, C) [0-255]
 
+
+		# Masking
+		model_template = get_model_options().get('template')
+		#model_size = get_model_options().get('size')
+		model_size = (int(r_x - l_x), int(r_y - l_y))
+		face_landmark_5 = target_face.landmark_set.get('5/68').copy()
+		crop_vision_frame, affine_matrix = warp_face_by_face_landmark_5(paste_vision_frame, face_landmark_5, model_template, model_size)
+		
+		box_mask = create_static_box_mask(crop_vision_frame.shape[:2], state_manager.get_item('face_mask_blur'), (0, 0, 0, 0))
+		#print("box_mask.shape", box_mask.shape)
+		cv2.imwrite("data/tests/mask_temp_vision_frame.jpg", temp_vision_frame)
+		cv2.imwrite("data/tests/mask_crop_vision_frame.jpg", crop_vision_frame)
+		cv2.imwrite("data/tests/mask_box_mask.jpg", box_mask*255.)
+		
+		crop_masks =\
+		[
+		box_mask
+		]
+
+		if 'occlusion' in state_manager.get_item('face_mask_types'):
+			occlusion_mask = create_occlusion_mask(crop_vision_frame)
+			cv2.imwrite("data/tests/mask_occlusion_crop_vision_frame.jpg", occlusion_mask*255.)
+			crop_masks.append(occlusion_mask)
+		if 'region' in state_manager.get_item('face_mask_types'):
+			region_mask = create_region_mask(crop_vision_frame, state_manager.get_item('face_mask_regions'))
+			cv2.imwrite("data/tests/mask_region_mask.jpg", region_mask*255.)
+			crop_masks.append(region_mask)
+		
+		crop_mask = np.minimum.reduce(crop_masks).clip(0, 1)
+		
+		cv2.imwrite("data/tests/mask_reduced_crop_mask.jpg", crop_mask*255.)
+
+		paste_vision_frame = paste_back(temp_vision_frame, crop_vision_frame, crop_mask, affine_matrix)
+		cv2.imwrite("data/tests/result.jpg", paste_vision_frame)
+		
+
+		# show rectangle to see the cropping rectangle (after the aging process)
+		if state_manager.get_item('age_modifier_show_mask') == "Yes":
+			cv2.rectangle(paste_vision_frame, (l_x, l_y), (r_x, r_y), (0, 0, 255), 3)
+
 		# show rectangle to see the cropping rectangle (after the aging process)
 		if state_manager.get_item('age_modifier_show_mask') == "Yes":
 			cv2.rectangle(paste_vision_frame, (l_x, l_y), (r_x, r_y), (0, 0, 255), 3)
